BDOG-523 Drop SessionStore and move cache.collection.PlayMongoCacheCollection to cache.MongoCacheRepository
