+++ hmrc-mongo-play-26/src/main/scala/uk/gov/hmrc/mongo/cache/model.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.mongo.cache.collection
+package uk.gov.hmrc.mongo.cache
 
 import java.time.Instant
 import play.api.libs.json.JsObject
 
-final case class CacheItem(
+private [cache] final case class CacheItem(
   id: String,
   data: JsObject,
   createdAt: Instant,
